[refactor] 收藏 API 的 callback 函式整理為 favoritesController.js
